requirements.txt
Babel
camelcase
certifi
charset-normalizer
Django
django-admin-decorators
django-admin-list-filter-dropdown
django-ckeditor
django-ckeditor-5
django-crispy-forms
django-filter
django-js-asset
django-render-partial
djangorestframework
djangorestframework-simplejwt
django-humanize
fonttools
jmespath
jsonfield
kavenegar
Khayyam
mysql-connector-python
Pillow
pipreqs
python-bsonjs
python-dateutil
python-dispatch
pytz
requests
six
sqlparse
tzdata
urllib3
virtualenv
captcha
django-simple-captcha
humanize
django-jazzmin
crispy-bootstrap5
django-jalali-date 